pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cloud-companion"
version = "0.1.0"
description = "AI-powered cloud resource analysis and troubleshooting chatbot"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [{name = "Cloud Companion Contributors", email = "info@cloudcompanion.dev"}]
keywords = ["cloud", "ai", "chatbot", "aws", "azure", "gcp", "troubleshooting"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: System :: Monitoring",
]

dependencies = [
    "fastapi==0.128.4",
    "uvicorn[standard]==0.40.0",
    "neo4j==6.1.0",
    "weaviate-client==4.19.2",
    "litellm==1.13.0",
    "boto3==1.29.7",
    "azure-identity==1.14.0",
    "azure-mgmt-compute==30.6.0",
    "google-cloud-compute==1.13.0",
    "httpx==0.25.2",
    "bcrypt==4.1.1",
    "pyjwt==2.8.1",
]

[project.optional-dependencies]
dev = [
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "pytest-cov==7.0.0",
    "ruff==0.15.0",
    "black==26.1.0",
    "mypy==1.19.1",
    "pre-commit==4.5.1",
    "ipython==8.38.0",
]

docs = [
    "mkdocs==1.6.1",
    "mkdocs-material==9.7.1",
]

[project.urls]
Homepage = "https://github.com/dm669230/cloud-companion"
Documentation = "https://cloud-companion.readthedocs.io"
Repository = "https://github.com/dm669230/cloud-companion.git"
Issues = "https://github.com/dm669230/cloud-companion/issues"
Changelog = "https://github.com/dm669230/cloud-companion/releases"

[tool.setuptools]
packages = ["app"]

[tool.black]
line-length = 100
target-version = ["py310"]
extend-exclude = '''
/(
  \.git
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "C",    # flake8-comprehensions
    "B",    # flake8-bugbear
]
ignore = ["E501"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=app --cov-report=html --cov-report=term-missing"
